backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
pydantic
slowapi